March 7, 2022 11:23 PM

+++ Day 02/Day 02-2ex/main.py
@@ -4,6 +4,6 @@ weight = input("enter your weight in kg: ")
 # 🚨 Don't change the code above 👆
 
 #Write your code below this line 👇
-BMI = int(height) / (int (weight) ** 2)
+BMI = float(height) / (float (weight) ** 2)
 
 print(BMI)
